mozilla/browser/fuel/test/test_Application.html

<html>
<head>
  <title>Testing Application.js</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"></link>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

test_Application();

function test_Application() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

  ok(Application, "Check global access to Application");
}

</script>
</pre>
</body>
</html>
